refactor ale module to be smaller
